demo3-react/inventory-service/requirements.txt
fastapi
uvicorn
motor
psycopg2-binary
opentelemetry-distro
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-pymongo
opentelemetry-exporter-otlp
python-dotenv